switched from useEffect to Loader in main and added a pages folder as well as Home and BookPage files
