+++ client/src/main.jsx
@@ -4,11 +4,22 @@ import ReactDOM from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
 import App from "./App";
+import Home from "./pages/Home";
+
+import GetBook from "./GetBook";
 
 const router = createBrowserRouter([
   {
-    path: "/",
     element: <App />,
+    id: "app",
+    children: [
+      {
+        path: "/",
+        id: "home",
+        element: <Home />,
+        loader: () => GetBook(),
+      },
+    ],
   },
 ]);
